# CraftWineShop/openAIgenerate/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAI API
  description: OpenAI API
  contact:
    name: Ihor Murashko
    email: i.murashko0911@gmail.com
  version: "1.0"
servers:
- url: https://craft-wine-shop.onrender.com
...
paths:
...
  /api/v1/user/get_user:
    get:
      tags:
      - user-personal-controller
      description: отримати користувача
      operationId: getUserInformation
      parameters:
      - name: Authorization
        in: header
        description: токен в заголовці Authorization
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ0aGlzQGVtYWlsLmNvbSIsInJvbGUiOiJVU0VSIiwiaWQiOjEsImV4cCI6MTcxMDYxNzE5NiwiaWF0IjoxNzEwNTMwNzk2LCJlbWFpbCI6InRoaXNAZW1haWwuY29tIn0.xQViqf8XnVGuQjcdK9tSxgCbb0kMvZdXv48-PD4Z7TeGFVcSrkdnV4MDGB8yshCVJypCZR0u-na_dSJARl7Z_w
      responses:
        "200":
          description: інформація про користувача з отриманого токену
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
...
components:
  schemas:
...
    CraftWineDTOResponse:
      type: object
      properties:
        id:
          type: integer
          description: унікальний цифровий ідентифікатор вина
          format: int64
          example: 5
        wineName:
          type: string
          description: ім'я
        price:
          type: number
          description: актуальна ціна вина з урахуванням поточних діючих знижок (акцій)
          example: 184.85
        adminDiscountPercentage:
          type: number
          description: знижка від адміністратора (у відсотках)
          format: float
          example: 22.5
        wineDescription:
          type: string
          description: опис вина
        quantity:
          type: integer
          description: залишок бутилок в магазині
          format: int32
        bottleCapacity:
          type: string
          description: об'єм бутилки
          example: "0.7"
        alcohol:
          type: string
          description: відсоток алкоголю
          example: "7.5"
        isNewCollection:
          type: boolean
          description: чи відноситься вино до нової колекції
        isBestSeller:
          type: boolean
          description: чи відноситься вино до бестселлерів
        isSale:
          type: boolean
          description: чи є вино зі знижкою
        isWineTimePromotion:
          type: boolean
          description: 'чи діє зараз акція "час вина" на даний товар '
        winemaking:
          type: string
        grapeVarieties:
          type: string
        tastingNotes:
          type: string
        storeAndServeAdvices:
          type: string
        foodPairing:
          type: string
        reviewsAndAwards:
          type: string
        wineColor:
          type: string
        sugarConsistency:
          type: string
        country:
          $ref: '#/components/schemas/ProducedCountryDTO'
        region:
          $ref: '#/components/schemas/RegionDTO'
        evaluation:
          type: integer
          description: може містити тільки ціле число від нуля (якщо ще жодної оцінки
            не було) до п'яти
          format: int32
          example: 4
        wineComments:
          type: array
          description: "список коментарів для данного вина відсортированих за дато\
            ю додавання (від більш нових, до старіших))"
          items:
            $ref: '#/components/schemas/WineCommentResponseDTO'
        bottlesSoldCounter:
          type: integer
          description: "ціле число, що репрезентує кількість проданих бутилок"
          format: int64
          example: 11
        addedDateTime:
          type: string
          description: дата та час коли товар був доданий у магазин
          format: date-time
        imageUrl:
          type: string
          description: посилання на зображення вина
          example: https://res.cloudinary.com/dtfnirefg/image/upload/v1707609852/craft-wine-shop/wine-2.png
      description: репрезентація сутності вина
    ProducedCountryDTO:
      type: object
      properties:
        id:
          type: integer
          description: унікальний ідентифікатор країни
          format: int64
        name:
          type: string
          description: Ім'я країни
      description: Країна виробник вина
    RegionDTO:
      type: object
      properties:
        id:
          type: integer
          description: унікальний ідентифікатор регіону
          format: int64
        name:
          type: string
          description: Ім'я регіону
      description: Регіон виробництва вина
    UserDTO:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        deliveryAddress:
          type: string
        favorites:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/CraftWineDTOResponse'
    WineCommentResponseDTO:
      type: object
      properties:
        id:
          type: integer
          description: унікальний ідентифікатор коментаря
          format: int64
        userFirstName:
          type: string
          description: "ім'я користувача, який залишив коментар"
        userLastName:
          type: string
          description: "прізвище користувача, який залишив коментар"
        comment:
          type: string
          description: коментар
        addedCommentTime:
          type: string
          description: дата додавання коментаря
          format: date-time
      description: Сутність для кожного коментаря
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT